view implementada e finalizado
