+++ src/app/app.component.html
@@ -1,26 +1,33 @@
 <h1>Accessible Modal Component</h1>
 
 <p>Welcome to the accessible modal testing page!</p>
-<button class="button button-primary" type="button" (click)="show()">Show modal</button>
+<button
+  class="button button-primary"
+  type="button"
+  (click)="show()">Show modal</button>
 
 <ng-template #modal>
   <div class="form-field">
     <label for="firstName">First name</label>
-    <input id="firstName" [(value)]="firstName" />
+    <input id="firstName" [(value)]="firstName">
   </div>
 
   <div class="form-field">
     <label for="surname">Surname</label>
-    <input id="surname" />
+    <input id="surname">
   </div>
 
   <div class="form-field">
     <label for="age">Age</label>
-    <input id="age" />
+    <input id="age">
   </div>
 
   <div class="button-group">
     <button class="button button-primary" type="button">Submit</button>
-    <button class="button button-secondary" type="button" (click)="modalRef.close()">Cancel</button>
+    <button
+      class="button button-secondary"
+      type="button"
+      (click)="modalRef.close()">Cancel
+    </button>
   </div>
 </ng-template>
