+++ src/app/shared/components/modal/modal.component.html
@@ -1,7 +1,6 @@
 <div class="modal-overlay">
   <div class="modal-content">
     <h1 class="modal-title">{{ config.title }}</h1>
-    <!-- conteudo do modal -->
     <ng-container *ngTemplateOutlet="config.templateRef"></ng-container>
   </div>
 </div>
