+++ src/app/shared/components/modal/modal.component.ts
@@ -1,10 +1,10 @@
-import { Component, TemplateRef } from '@angular/core';
+import { Component } from '@angular/core';
 import { ModalConfig } from './interfaces/modal-config';
 
 @Component({
   selector: 'app-modal',
   templateUrl: './modal.component.html',
-  styleUrls: ['modal.component.scss'],
+  styleUrls: ['modal.component.scss']
 })
 export class ModalComponent {
   public config: ModalConfig;
